sprint: fix M-022 M-023 H-030 M-031 M-032 M-033 M-034
- net: tcp_abort_socket() frees a kernel-owned child; used on retry
  exhaustion (H-030) and when the listener is gone at the final ACK (M-023)
- socket: accept() sets the child's owner (M-031); the peer-address pointer
  is checked before a connection/datagram is consumed (M-032)
- fd: size/pos are 64-bit so fstat/lseek do not wrap past 4 GiB (M-033)
- ext2: directory block count without the overflowing sum, clamped (M-022)
- boot: ELF/PE entry must lie inside an executable segment (M-034)

Red tests first for each (verified red with the fixes removed); check
VERDICT was green apart from the chokepoint count, raised 44->46 for M-032.

+++ kernel/arch/x86_64/arch_hw_internal.h
@@ -154,8 +154,8 @@ typedef struct {
      * `writable` then says which end. */
     int pipe;
     uint32_t cluster;
-    uint32_t size;
-    uint32_t pos;
+    uint64_t size;        /* 64-bit: a >4 GiB file must not wrap in fstat/lseek (M-033) */
+    uint64_t pos;
     int net_sock;         /* index into the TCP/IP stack, or -1 for a file */
     uint32_t dir_index;   /* for getdents64 on a directory fd */
     /* Whether this descriptor names a directory. Determined when it is opened
